Refactoring: RegressionTask trainer
