+++ tests/trainers/test_cyclone.py
@@ -2,15 +2,10 @@
 # Licensed under the MIT License.
 
 import os
-from typing import Any, Dict, Generator, cast
 
 import pytest
-from _pytest.monkeypatch import MonkeyPatch
-from omegaconf import OmegaConf
 
-from torchgeo.trainers import CycloneDataModule, CycloneSimpleRegressionTask
-
-from .test_utils import mocked_log
+from torchgeo.trainers import CycloneDataModule
 
 
 @pytest.fixture(scope="module")
@@ -23,54 +18,6 @@ def datamodule() -> CycloneDataModule:
     dm.prepare_data()
     dm.setup()
     return dm
-
-
-class TestCycloneSimpleRegressionTask:
-    @pytest.fixture
-    def config(self) -> Dict[str, Any]:
-        task_conf = OmegaConf.load(
-            os.path.join("conf", "task_defaults", "cyclone.yaml")
-        )
-        task_args = OmegaConf.to_object(task_conf.experiment.module)
-        task_args = cast(Dict[str, Any], task_args)
-        return task_args
-
-    @pytest.fixture
-    def task(
-        self, config: Dict[str, Any], monkeypatch: Generator[MonkeyPatch, None, None]
-    ) -> CycloneSimpleRegressionTask:
-        task = CycloneSimpleRegressionTask(**config)
-        monkeypatch.setattr(task, "log", mocked_log)  # type: ignore[attr-defined]
-        return task
-
-    def test_configure_optimizers(self, task: CycloneSimpleRegressionTask) -> None:
-        out = task.configure_optimizers()
-        assert "optimizer" in out
-        assert "lr_scheduler" in out
-
-    def test_training(
-        self, datamodule: CycloneDataModule, task: CycloneSimpleRegressionTask
-    ) -> None:
-        batch = next(iter(datamodule.train_dataloader()))
-        task.training_step(batch, 0)
-
-    def test_validation(
-        self, datamodule: CycloneDataModule, task: CycloneSimpleRegressionTask
-    ) -> None:
-        batch = next(iter(datamodule.val_dataloader()))
-        task.validation_step(batch, 0)
-
-    def test_test(
-        self, datamodule: CycloneDataModule, task: CycloneSimpleRegressionTask
-    ) -> None:
-        batch = next(iter(datamodule.test_dataloader()))
-        task.test_step(batch, 0)
-
-    def test_invalid_model(self, config: Dict[str, Any]) -> None:
-        config["model"] = "invalid_model"
-        error_message = "Model type 'invalid_model' is not valid."
-        with pytest.raises(ValueError, match=error_message):
-            CycloneSimpleRegressionTask(**config)
 
 
 class TestCycloneDataModule:
