+++ tests/trainers/test_tasks.py
@@ -9,7 +9,12 @@ from _pytest.fixtures import SubRequest
 from _pytest.monkeypatch import MonkeyPatch
 from omegaconf import OmegaConf
 
-from torchgeo.trainers import ClassificationTask, So2SatDataModule
+from torchgeo.trainers import (
+    ClassificationTask,
+    So2SatDataModule,
+    RegressionTask,
+    CycloneDataModule,
+)
 
 from .test_utils import mocked_log
 
@@ -110,3 +115,60 @@ class TestClassificationTask:
         error_message = "Trying to load resnet18 weights into a resnet50"
         with pytest.raises(ValueError, match=error_message):
             ClassificationTask(**config)
+
+
+class TestRegressionTask:
+    @pytest.fixture(scope="class")
+    def datamodule(self) -> CycloneDataModule:
+        root = os.path.join("tests", "data", "cyclone")
+        seed = 0
+        batch_size = 1
+        num_workers = 0
+        dm = CycloneDataModule(root, seed, batch_size, num_workers)
+        dm.prepare_data()
+        dm.setup()
+        return dm
+
+    @pytest.fixture
+    def config(self) -> Dict[str, Any]:
+        task_conf = OmegaConf.load(
+            os.path.join("conf", "task_defaults", "cyclone.yaml")
+        )
+        task_args = OmegaConf.to_object(task_conf.experiment.module)
+        task_args = cast(Dict[str, Any], task_args)
+        return task_args
+
+    @pytest.fixture
+    def task(
+        self, config: Dict[str, Any], monkeypatch: Generator[MonkeyPatch, None, None]
+    ) -> RegressionTask:
+        task = RegressionTask(**config)
+        monkeypatch.setattr(task, "log", mocked_log)  # type: ignore[attr-defined]
+        return task
+
+    def test_configure_optimizers(self, task: RegressionTask) -> None:
+        out = task.configure_optimizers()
+        assert "optimizer" in out
+        assert "lr_scheduler" in out
+
+    def test_training(
+        self, datamodule: CycloneDataModule, task: RegressionTask
+    ) -> None:
+        batch = next(iter(datamodule.train_dataloader()))
+        task.training_step(batch, 0)
+
+    def test_validation(
+        self, datamodule: CycloneDataModule, task: RegressionTask
+    ) -> None:
+        batch = next(iter(datamodule.val_dataloader()))
+        task.validation_step(batch, 0)
+
+    def test_test(self, datamodule: CycloneDataModule, task: RegressionTask) -> None:
+        batch = next(iter(datamodule.test_dataloader()))
+        task.test_step(batch, 0)
+
+    def test_invalid_model(self, config: Dict[str, Any]) -> None:
+        config["model"] = "invalid_model"
+        error_message = "Model type 'invalid_model' is not valid."
+        with pytest.raises(ValueError, match=error_message):
+            RegressionTask(**config)
