+++ torchgeo/trainers/__init__.py
@@ -5,24 +5,24 @@
 
 from .byol import BYOLTask
 from .chesapeake import ChesapeakeCVPRDataModule, ChesapeakeCVPRSegmentationTask
-from .cyclone import CycloneDataModule, CycloneSimpleRegressionTask
+from .cyclone import CycloneDataModule
 from .landcoverai import LandcoverAIDataModule, LandcoverAISegmentationTask
 from .naipchesapeake import NAIPChesapeakeDataModule, NAIPChesapeakeSegmentationTask
 from .resisc45 import RESISC45ClassificationTask, RESISC45DataModule
 from .sen12ms import SEN12MSDataModule, SEN12MSSegmentationTask
 from .so2sat import So2SatClassificationTask, So2SatDataModule
-from .tasks import ClassificationTask
+from .tasks import ClassificationTask, RegressionTask
 from .ucmerced import UCMercedClassificationTask, UCMercedDataModule
 
 __all__ = (
     # Tasks
     "ClassificationTask",
+    "RegressionTask",
     # Trainers
     "BYOLTask",
     "ChesapeakeCVPRSegmentationTask",
     "ChesapeakeCVPRDataModule",
     "CycloneDataModule",
-    "CycloneSimpleRegressionTask",
     "LandcoverAIDataModule",
     "LandcoverAISegmentationTask",
     "NAIPChesapeakeDataModule",
