+++ train.py
@@ -18,11 +18,11 @@ from torchgeo.trainers import (
     ChesapeakeCVPRDataModule,
     ChesapeakeCVPRSegmentationTask,
     CycloneDataModule,
-    CycloneSimpleRegressionTask,
     LandcoverAIDataModule,
     LandcoverAISegmentationTask,
     NAIPChesapeakeDataModule,
     NAIPChesapeakeSegmentationTask,
+    RegressionTask,
     RESISC45ClassificationTask,
     RESISC45DataModule,
     SEN12MSDataModule,
@@ -38,7 +38,7 @@ TASK_TO_MODULES_MAPPING: Dict[
 ] = {
     "byol": (BYOLTask, ChesapeakeCVPRDataModule),
     "chesapeake_cvpr": (ChesapeakeCVPRSegmentationTask, ChesapeakeCVPRDataModule),
-    "cyclone": (CycloneSimpleRegressionTask, CycloneDataModule),
+    "cyclone": (RegressionTask, CycloneDataModule),
     "landcoverai": (LandcoverAISegmentationTask, LandcoverAIDataModule),
     "naipchesapeake": (NAIPChesapeakeSegmentationTask, NAIPChesapeakeDataModule),
     "resisc45": (RESISC45ClassificationTask, RESISC45DataModule),
